Refactor the entire solution.
This patch almost create everything from the scratch. The parser
is only responsible for parsing and passing each found token to
the handler which makes decision what should be done with it.
Therefore it was very easy to implement JSON compacting. The patch
introduces command line parsing for determining which command and
options were requested by the caller.